Fix GluonTS tests
