+++ timeseries/tests/unittests/models/gluonts/mx/test_mx.py
@@ -1,5 +1,4 @@
 from functools import partial
-from unittest import mock
 
 import pytest
 
@@ -19,9 +18,8 @@ from autogluon.timeseries.models.gluonts.mx import (
 )
 from autogluon.timeseries.models.gluonts.mx.models import GenericGluonTSMXNetModelFactory
 from autogluon.timeseries.models.presets import get_default_hps
-from autogluon.timeseries.utils.features import ContinuousAndCategoricalFeatureGenerator
 
-from ....common import DATAFRAME_WITH_STATIC, DUMMY_TS_DATAFRAME
+from ....common import DUMMY_TS_DATAFRAME
 
 TESTABLE_MX_MODELS = [
     DeepARMXNetModel,
@@ -89,25 +87,3 @@ def test_when_tft_quantiles_are_deciles_then_forecast_contains_correct_quantiles
 def test_when_mxnet_installed_then_default_presets_include_mxnet_models(preset_key):
     hps = get_default_hps(key=preset_key, prediction_length=5)
     assert any("MXNet" in model_name for model_name in hps.keys())
-
-
-@pytest.fixture(scope="module")
-def df_with_static():
-    feature_pipeline = ContinuousAndCategoricalFeatureGenerator()
-    df = DATAFRAME_WITH_STATIC.copy(deep=False)
-    df.static_features = feature_pipeline.fit_transform(df.static_features)
-    return df
-
-
-@pytest.mark.parametrize("model_class", TESTABLE_MX_MODELS_WITH_STATIC_FEATURES)
-def test_given_fit_with_static_features_when_predicting_then_static_features_are_used(model_class, df_with_static):
-    model = model_class(hyperparameters={"epochs": 1, "num_batches_per_epoch": 1})
-    model.fit(train_data=df_with_static)
-    with mock.patch("gluonts.mx.model.predictor.RepresentableBlockPredictor.predict") as mock_predict:
-        try:
-            model.predict(df_with_static)
-        except IndexError:
-            gluonts_dataset = mock_predict.call_args[1]["dataset"]
-            item = next(iter(gluonts_dataset))
-            assert item["feat_static_cat"].shape == (1,)
-            assert item["feat_static_real"].shape == (2,)
